refactor: 공부계획/시험공부계획 컴포넌트 폴더 및 파일명 수정
