+++ src/components/examplan/ExamPlanStep3.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const ExamPlanStep3 = ({
+  formData,
+  updateFormData,
+  prevStep,
+}) => {};
+
+export default ExamPlanStep3;
