+++ src/components/studyPlan/StudyPlanStep5.jsx
@@ -2,31 +2,30 @@ import React, { useState } from "react";
 import { ChevronRight } from "lucide-react";
 import bookLogo from "../../assets/logo/book_square.svg";
 
-const PlanStep3 = ({
+const StudyPlanStep5 = ({
   nextStep,
   updateFormData,
   formData = {},
   handleClose,
 }) => {
   const options = [
-    "5:5",
-    "6:4",
-    "7:3",
-    "8:2",
-    "9:1",
+    "바로 시작",
+    "마감 하루 전",
+    "마감 3일 전",
+    "마감 1주 전",
     "직접 입력",
   ];
 
   const [selectedOption, setSelectedOption] = useState(
-    formData.studyRestRatio &&
-      !options.includes(formData.studyRestRatio)
+    formData.assignmentTiming &&
+      !options.includes(formData.assignmentTiming)
       ? "직접 입력"
-      : formData.studyRestRatio || ""
+      : formData.assignmentTiming || ""
   );
   const [customValue, setCustomValue] = useState(
-    options.includes(formData.studyRestRatio)
+    options.includes(formData.assignmentTiming)
       ? ""
-      : formData.studyRestRatio || ""
+      : formData.assignmentTiming || ""
   );
 
   const handleSelect = (option) => {
@@ -49,33 +48,50 @@ const PlanStep3 = ({
 
     let structuredValue;
 
-    if (valueToSubmit.match(/^\d+:\d+$/)) {
-      const [study, rest] = valueToSubmit.split(":").map(Number);
-      const total = study + rest;
-
-      if (total === 0) {
-        alert("유효한 비율을 입력해주세요.");
-        return;
+    if (selectedOption === "바로 시작") {
+      structuredValue = { type: "immediate" };
+    } else if (selectedOption.startsWith("마감 ")) {
+      const match = selectedOption.match(
+        /마감\s(\d+)(일|주)\s전/
+      );
+      if (match) {
+        const amount = parseInt(match[1]);
+        const unit = match[2] === "일" ? "day" : "week";
+        structuredValue = {
+          type: "relative",
+          offset: -amount,
+          unit,
+          reference: "deadline",
+        };
       }
-
-      structuredValue = {
-        type: "ratio",
-        studyPercent: Math.round((study / total) * 100),
-        restPercent: Math.round((rest / total) * 100),
-      };
-    } else {
-      structuredValue = {
-        type: "custom",
-        raw: valueToSubmit,
-      };
+    } else if (selectedOption === "직접 입력") {
+      const match = customValue.match(
+        /마감\s?(\d+)(일|주)\s?전/
+      );
+      if (match) {
+        const amount = parseInt(match[1]);
+        const unit = match[2] === "일" ? "day" : "week";
+        structuredValue = {
+          type: "relative",
+          offset: -amount,
+          unit,
+          reference: "deadline",
+          raw: customValue,
+        };
+      } else {
+        structuredValue = {
+          type: "custom",
+          raw: customValue,
+        };
+      }
     }
 
-    console.log("사용자 선택값:", valueToSubmit);
-    console.log("구조화된 값:", structuredValue);
+    console.log("선택한 값:", valueToSubmit);
+    console.log("구조화된 데이터:", structuredValue);
 
     updateFormData({
-      studyRestRatio: valueToSubmit,
-      studyRestRatioStructured: structuredValue,
+      assignmentTiming: valueToSubmit,
+      assignmentTimingStructured: structuredValue,
     });
     nextStep();
   };
@@ -111,7 +127,7 @@ const PlanStep3 = ({
           <div className="px-8 py-16 relative h-[calc(100%-110px)]">
             <div className="absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 flex flex-col items-center space-y-12">
               <p className="text-center text-gray-700 text-xl">
-                3. 공부와 휴식의 비율은 어떻게 되시나요?
+                5. 과제는 언제 시작/마감하는 것을 선호하시나요?
               </p>
 
               <div className="relative">
@@ -138,7 +154,7 @@ const PlanStep3 = ({
                     <input
                       type="text"
                       className="border-2 border-gray-300 rounded-lg px-6 py-4 w-80 text-center text-lg mt-8 focus:outline-none focus:border-[#9DB2BF] transition-colors"
-                      placeholder="직접 비율을 입력하세요"
+                      placeholder="예: 마감 2일 전, 주말 등"
                       value={customValue}
                       onChange={(e) =>
                         setCustomValue(e.target.value)
@@ -161,4 +177,4 @@ const PlanStep3 = ({
   );
 };
 
-export default PlanStep3;
+export default StudyPlanStep5;
