+++ src/router/AppRouter.jsx
@@ -5,8 +5,8 @@ import Timetable from "../pages/Timetable";
 import ChatbotComponent from "../components/chatbot/ChatbotComponent";
 import Loginpage from "../pages/Login";
 import MainLayout from "../components/layout/MainLayout";
-import StudyPlanStep from "../components/studyPlan/StudyPlanStep";
-import PlanStep from "../components/planSetup/PlanStep";
+import ExamPlanStep from "../components/examplan/ExamPlanStep";
+import StudyPlanStep from "../components/studyplan/StudyPlanStep";
 
 //setting에 사용되는 컴포넌트들
 import Profile from "../components/settingsDetail/Profile";
@@ -33,9 +33,9 @@ const AppRouter = () => {
         </Route>
         <Route path="/timetable" element={<Timetable />} />
         <Route path="/chatbot" element={<ChatbotComponent />} />
-        <Route path="/study-plan" element={<StudyPlanStep />} />
+        <Route path="/examplan" element={<ExamPlanStep />} />
       </Route>
-      <Route path="/plan-setup" element={<PlanStep />} />
+      <Route path="/studyplan/setup" element={<StudyPlanStep />} />
       <Route path="/login" element={<Loginpage />} />
       <Route path="/signup" element={<Signup />} />
     </Routes>
